# uni_core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  request.DockerRegistrySaveReq:
    properties:
      password:
        description: Password 是仓库密码（明文仅在本次请求内瞬时存在；**每次创建/更新都重输**）。
        type: string
      registry:
        description: |-
          Registry 是仓库地址（唯一键：主机名/IP + 可选端口，不含协议头与镜像路径；
          形态校验在 service 复用协议 IsDockerRegistryAddr）。
        type: string
      remark:
        description: Remark 是备注（可选）。
        type: string
      username:
        description: Username 是仓库登录名。
        type: string
    required:
    - password
    - registry
    - username
    type: object
...
  response.DockerRegistryItem:
    properties:
      createdAt:
        description: CreatedAt 是创建时刻（unix 秒）。
        type: integer
      password:
        description: Password 恒为掩码「****」。
        type: string
      registry:
        description: Registry 是仓库地址（唯一键；不含协议头与镜像路径）。
        type: string
      remark:
        description: Remark 是备注。
        type: string
      username:
        description: Username 是仓库登录名。
        type: string
    type: object
  response.DockerRegistryListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/response.DockerRegistryItem'
        type: array
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: 企业级 RBAC 权限管理系统后端 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RBAC Admin API
  version: 1.0.0
paths:
...
  /docker/events:
    get:
      description: 跨主机聚合的容器/镜像/卷/网络事件流；连接后先回放各主机最近 50 条，再实时推送；每行 {"host_id","hostname","t","type","action","actor_name","actor_id"}
      produces:
      - application/x-ndjson
      responses:
        "200":
          description: 流已建立(逐行 NDJSON)
        "401":
          description: 未登录
        "403":
          description: 无权限(docker:list)
      security:
      - BearerAuth: []
      summary: Docker 事件聚合流(NDJSON)
      tags:
      - Docker 管理
...
  /docker/registries:
    get:
      description: 枚举全部私有仓库凭据（registry/用户名/备注/掩码密码/创建时间；密码任何读路径都不回明文）
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerRegistryListResp'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 仓库凭据列表
      tags:
      - Docker 管理
    post:
      consumes:
      - application/json
      description: 新建私有仓库凭据（registry 唯一键；密码 AES-256-GCM 加密落库，一切读路径掩码）
      parameters:
      - description: 凭据请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.DockerRegistrySaveReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerRegistryItem'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
        "409":
          description: 该仓库地址已有凭据
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 新建仓库凭据
      tags:
      - Docker 管理
    put:
      consumes:
      - application/json
      description: 按 registry 更新凭据的用户名/密码/备注（密码必须重输，凭据管理没有读回旧密码的路径）
      parameters:
      - description: 凭据请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.DockerRegistrySaveReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.DockerRegistryItem'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 该仓库没有凭据记录
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 更新仓库凭据
      tags:
      - Docker 管理
  /docker/registries/{registry}:
    delete:
      description: 按仓库地址删除凭据（删除即失效，不带任何对账周期）
      parameters:
      - description: 仓库地址
        in: path
        name: registry
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: 无权限
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 该仓库没有凭据记录
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 删除仓库凭据
      tags:
      - Docker 管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"